backend/requirements.txt
flask==3.0.0
flask-cors
python-dotenv
pandas
openai
langchain==0.1.0
tqdm
chromadb
sentence-transformers
PyMuPDF         
langchain-community
pypdf
tiktoken
faiss-cpu
reportlab
pypdf2==3.0.1
flask-dance
flask-session
langchain_openai==0.0.2
spacy
python-magic-bin
python-docx==1.0.1